requirements.txt
streamlit
lightgbm
websockets
pandas
numpy
plotly
requests
websocket-client
